requirements.txt
Pillow==8.4.0
pytest==6.2.5
aiogram~=2.16
pytest-asyncio==0.16.0
python-dotenv==0.19.1
